some files are move for architecture of project
