--- src/main/java/com/uniquindio/reporte/entities/enums/users/EnumUserType.java
@@ -1,6 +1,0 @@
-package com.uniquindio.reporte.entities.enums.users;
-
-public enum EnumUserType {
-    ADMIN,
-    USER,
-}
