+++ src/main/java/com/uniquindio/reporte/model/entities/Category.java
@@ -1,6 +1,6 @@
-package com.uniquindio.reporte.entities;
+package com.uniquindio.reporte.model.entities;
 
-
+import com.uniquindio.reporte.model.entities.enums.reports.EnumCategoryReport;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import org.bson.types.ObjectId;
@@ -8,17 +8,20 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
 
-import java.time.LocalDateTime;
-
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @Data
-@Document(collection = "verification_email")
-public class VerificationEmail {
+@Document(collection = "category")
+public class Category {
 
     @EqualsAndHashCode.Include
     @Id
     ObjectId Id;
+    String name;
+    String description;
+    @Field(name="activation_status")
+    Boolean activationStatus;
     User user;
-    @Field(name="date_time")
-    LocalDateTime dateTime;
+    @Field(name="category_report")
+    EnumCategoryReport categoryReport;
+
 }
