+++ src/main/java/com/uniquindio/reporte/model/entities/Comment.java
@@ -1,4 +1,4 @@
-package com.uniquindio.reporte.entities;
+package com.uniquindio.reporte.model.entities;
 
 
 import lombok.Data;
