+++ src/main/java/com/uniquindio/reporte/model/entities/FileDocument.java
@@ -1,23 +1,26 @@
-package com.uniquindio.reporte.entities;
-
+package com.uniquindio.reporte.model.entities;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
+import org.springframework.data.mongodb.core.mapping.Field;
+
+import java.time.LocalDate;
 
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @Data
-@Document(collection = "ubication")
-public class Ubication {
+@Document(collection = "file_document")
+public class FileDocument {
 
     @EqualsAndHashCode.Include
     @Id
     ObjectId Id;
-    float latitude;
-    float longitude;
+    String url;
     String name;
     String description;
-
+    @Field(name="creation_date")
+    LocalDate creationDate;
+    User user;
 }
