+++ src/main/java/com/uniquindio/reporte/model/entities/Notification.java
@@ -1,6 +1,6 @@
-package com.uniquindio.reporte.entities;
+package com.uniquindio.reporte.model.entities;
 
-import com.uniquindio.reporte.entities.enums.reports.EnumCategoryReport;
+
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import org.bson.types.ObjectId;
@@ -8,20 +8,20 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
 
+import java.time.LocalDate;
+
+
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @Data
-@Document(collection = "category")
-public class Category {
+@Document(collection = "notification")
+public class Notification {
 
     @EqualsAndHashCode.Include
     @Id
     ObjectId Id;
-    String name;
-    String description;
-    @Field(name="activation_status")
-    Boolean activationStatus;
+    String message;
+    @Field(name="shipping_date")
+    LocalDate shippingDate;
     User user;
-    @Field(name="category_report")
-    EnumCategoryReport categoryReport;
-
+    Report report;
 }
