+++ src/main/java/com/uniquindio/reporte/model/entities/Ubication.java
@@ -1,4 +1,4 @@
-package com.uniquindio.reporte.entities;
+package com.uniquindio.reporte.model.entities;
 
 
 import lombok.Data;
@@ -6,22 +6,18 @@ import lombok.EqualsAndHashCode;
 import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
-import org.springframework.data.mongodb.core.mapping.Field;
-
-import java.time.LocalDate;
-
 
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @Data
-@Document(collection = "notification")
-public class Notification {
+@Document(collection = "ubication")
+public class Ubication {
 
     @EqualsAndHashCode.Include
     @Id
     ObjectId Id;
-    String message;
-    @Field(name="shipping_date")
-    LocalDate shippingDate;
-    User user;
-    Report report;
+    float latitude;
+    float longitude;
+    String name;
+    String description;
+
 }
