+++ src/main/java/com/uniquindio/reporte/model/entities/VerificationEmail.java
@@ -1,4 +1,5 @@
-package com.uniquindio.reporte.entities;
+package com.uniquindio.reporte.model.entities;
+
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -7,20 +8,17 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
 
-import java.time.LocalDate;
+import java.time.LocalDateTime;
 
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @Data
-@Document(collection = "file_document")
-public class FileDocument {
+@Document(collection = "verification_email")
+public class VerificationEmail {
 
     @EqualsAndHashCode.Include
     @Id
     ObjectId Id;
-    String url;
-    String name;
-    String description;
-    @Field(name="creation_date")
-    LocalDate creationDate;
     User user;
+    @Field(name="date_time")
+    LocalDateTime dateTime;
 }
