+++ src/main/java/com/uniquindio/reporte/model/entities/enums/users/EnumUserType.java
@@ -1,0 +1,6 @@
+package com.uniquindio.reporte.model.entities.enums.users;
+
+public enum EnumUserType {
+    ADMIN,
+    USER,
+}
