+++ src/main/java/com/uniquindio/reporte/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.uniquindio.reporte.repository;
+
+import com.uniquindio.reporte.model.entities.User;
+import org.bson.types.ObjectId;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends MongoRepository<User, ObjectId> {
+}
